+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Bump MSRV from `1.65` to `1.70`.
 - Bump `windows-sys` from `0.48.0` to `0.52.0`.
 - Expose `Egl` and `Glx` raw API functions on `Egl` and `Glx` displays.
+- Add `GLUTIN_WGL_OPENGL_DLL` environment variable to change OpenGL provider name with WGL.
 
 # Version 0.31.3